==> notebooks/01_import_merge.ipynb <==
# 1. Импорт библиотек и настройка

import pandas as pd
import numpy as np

DATA_PATH = 'C:/Users/kokontsev/Desktop/Sleep-Quality-Prediction/data/raw/'
OUTPUT_PATH = "C:/Users/kokontsev/Desktop/Sleep-Quality-Prediction/data/processed/"

pd.set_option("display.max_columns", 100)
pd.set_option("display.width", 120)

# 2. Загрузка данных

daily_activity = pd.read_csv(DATA_PATH + "dailyActivity_merged.csv")
daily_steps = pd.read_csv(DATA_PATH + "dailySteps_merged.csv")
daily_calories = pd.read_csv(DATA_PATH + "dailyCalories_merged.csv")
daily_intensities = pd.read_csv(DATA_PATH + "dailyIntensities_merged.csv")
sleep = pd.read_csv(DATA_PATH + "sleepDay_merged.csv")

print(daily_activity.shape)
print(sleep.shape)
sleep.head()

# 3. Приведение дат

daily_activity["ActivityDate"] = pd.to_datetime(daily_activity["ActivityDate"]).dt.date
daily_steps["ActivityDay"] = pd.to_datetime(daily_steps["ActivityDay"]).dt.date
daily_calories["ActivityDay"] = pd.to_datetime(daily_calories["ActivityDay"]).dt.date
daily_intensities["ActivityDay"] = pd.to_datetime(daily_intensities["ActivityDay"]).dt.date
sleep["SleepDay"] = pd.to_datetime(sleep["SleepDay"]).dt.date

# 4. Формирование дневной таблицы активности

activity = daily_activity[[
    "Id",
    "ActivityDate",
    "TotalSteps",
    "TotalDistance",
    "Calories",
    "VeryActiveMinutes",
    "FairlyActiveMinutes",
    "LightlyActiveMinutes",
    "SedentaryMinutes"
]].copy()

activity = activity.rename(columns={
    "Id": "user_id",
    "ActivityDate": "date"
})

# 5. Подготовка таблицы сна

**Выбираем столбцы**

sleep_df = sleep[[
    "Id",
    "SleepDay",
    "TotalSleepRecords",
    "TotalMinutesAsleep",
    "TotalTimeInBed"
]].copy()

**Приводим к формату**

sleep_df = sleep_df.rename(columns={"Id": "user_id", "SleepDay": "date"})

**Добавляем производную целевую переменную**

sleep_df["SleepEfficiency"] = (
    sleep_df["TotalMinutesAsleep"] / sleep_df["TotalTimeInBed"]
)

# 6. Объединение: активность → сон

**ВАЖНО: Сон ночью относится к предыдущему дню активности!!!**

sleep_df["activity_date"] = sleep_df["date"] - pd.Timedelta(days=1)

df = sleep_df.merge(
    activity,
    left_on=["user_id", "date"],
    right_on=["user_id", "date"],
    how="left",
    suffixes=("", "_activity")
)

df = df.rename(columns={"date": "sleep_date"})

print(df.shape)
df.head()

# 7. Сохранение результата

df.to_csv(OUTPUT_PATH + "sleep_activity_daily.csv", index=False)

# 8. Итоги ноубука
- единый user_id
- одна строка = одна ночь сна
- целевая переменная SleepEfficiency
- активность за предыдущий день
- сохранён processed-файл
